Break up and refactor
