+++ winc-rs/src/stack.rs
@@ -1,0 +1,10 @@
+pub mod sock_holder;
+pub mod socket_callbacks;
+/// Shared stack code for both sync and async clients
+pub mod stack_error;
+
+use crate::socket::Socket;
+pub use sock_holder::SockHolder;
+pub use socket_callbacks::ClientSocketOp;
+use socket_callbacks::Handle;
+pub use stack_error::StackError;
